Separate stages of the Compiler
