+++ src/frontend/flags.rs
@@ -1,4 +1,4 @@
-use crate::codegen::{ERR_STR, NOTE_STR};
+use crate::backend::codegen::{ERR_STR, NOTE_STR};
 use std::{
     collections::HashMap,
     fmt::{Debug, Display, Formatter},
