+++ src/main.rs
@@ -1,14 +1,9 @@
 #![allow(unused, unreachable_code)]
 
-mod parser;
-mod checker;
-mod codegen;
-mod assembler;
+mod frontend;
+mod middleend;
+mod backend;
 mod compiler;
-
-mod nodes;
-mod flags;
-mod instr;
 
 use std::time::Instant;
 
@@ -24,7 +19,8 @@ fn main() {
 mod tests {
     use lazy_static::lazy_static;
 
-    use crate::{compiler::Compiler, codegen::ERR_STR};
+    use crate::compiler::Compiler;
+    use crate::backend::codegen::ERR_STR;
 
     const ALWAYS_FAILS: &str = "This is a String we defined to make sure that a test always fails. This is expected.";
 
@@ -77,7 +73,7 @@ mod tests {
 
     mod syntax_tests {
         use crate::compiler::Compiler;
-        use crate::codegen::ERR_STR;
+        use crate::backend::codegen::ERR_STR;
 
         macro_rules! generate_failing_test {
             ($name:ident, $($err:expr),*) => {
@@ -108,7 +104,8 @@ mod tests {
     }
 
     mod semantic_tests {
-        use crate::{compiler::Compiler, codegen::{ERR_STR}};
+        use crate::compiler::Compiler;
+        use crate::backend::codegen::ERR_STR;
         use crate::tests::ALWAYS_FAILS;
 
         macro_rules! generate_failing_test {
