+++ src/middleend/checker.rs
@@ -1,9 +1,10 @@
 use std::collections::{HashMap, VecDeque};
 use std::fmt::{Display, Formatter};
 
-use crate::{parser::Location, codegen::{ERR_STR, NOTE_STR, WARN_STR}};
+use crate::frontend::parser::Location;
+use crate::frontend::nodes;
 
-use crate::nodes;
+use crate::backend::codegen::{ERR_STR, NOTE_STR, WARN_STR};
 
 #[derive(Debug, PartialEq, Eq, Clone, Default)]
 pub enum Type {
